+++ src/components/Home.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import './Home.css';
+
 
 function Home() {
   const [tasks, setTasks] = useState([]);
