+++ src/components/Login.jsx
@@ -1,18 +1,17 @@
 import React from 'react';
-import './SignUp.css';
 
-function SignUp() {
+
+function Login() {
   return (
     <div className="container">
-      <h2>Sign Up</h2>
+      <h2>Login</h2>
       <form>
-        <input type="text" placeholder="Username" required />
         <input type="email" placeholder="Email" required />
         <input type="password" placeholder="Password" required />
-        <button type="submit">Sign Up</button>
+        <button type="submit">Login</button>
       </form>
     </div>
   );
 }
 
-export default SignUp;
+export default Login;
