+++ src/components/SignUp.jsx
@@ -1,17 +1,18 @@
 import React from 'react';
-import './Login.css';
 
-function Login() {
+
+function SignUp() {
   return (
     <div className="container">
-      <h2>Login</h2>
+      <h2>Sign Up</h2>
       <form>
+        <input type="text" placeholder="Username" required />
         <input type="email" placeholder="Email" required />
         <input type="password" placeholder="Password" required />
-        <button type="submit">Login</button>
+        <button type="submit">Sign Up</button>
       </form>
     </div>
   );
 }
 
-export default Login;
+export default SignUp;
